Organized custom inspector features to match MRTK patterns.
