+++ Assets/MixedRealityToolkit-SDK/Features/UX/Interactable/Scripts/Events/InteractableOnToggleReceiver.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using Microsoft.MixedReality.Toolkit.Core.Utilities.InspectorFields;
+using System.Collections;
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
