+++ Assets/MixedRealityToolkit-SDK/Features/UX/Interactable/Scripts/Profile/InteractableProfileItem.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Microsoft.MixedReality.Toolkit.Core.Utilities.InspectorFields;
 using System;
 using System.Collections;
 using System.Collections.Generic;
