+++ Assets/MixedRealityToolkit/_Core/Inspectors/Utilities/InspectorFieldsUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3185493363fa80347a2ec0f9fa8907db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
