+++ Assets/MixedRealityToolkit/_Core/Inspectors/Utilities/InspectorUIUtility.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: b36edd39d49efe1468094b721f2d37f6
-timeCreated: 1520548123
-licenseType: Pro
+guid: 8ef3c8df76e6d8b44ad3a0fb3c3a7cc3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
