+++ Assets/MixedRealityToolkit/_Core/Utilities/InspectorFields.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e680f6084c4c8f43bbf4ec60b4288af
+guid: 793c3f0fd5016dc419f48fef64e2c0d7
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
